requirements/requirements.txt
# Core Flask
Flask==3.1.0
Werkzeug==3.1.3
flask-cors==4.0.0

# TensorFlow (includes built-in Keras)
tensorflow==2.15.0
h5py==3.9.0
numpy==1.24.3

# Image Processing
opencv-python==4.8.1.78
Pillow==9.5.0

# Data Processing
pandas==1.5.3
scikit-learn==1.3.2
scipy==1.10.1
statsmodels==0.14.0

# ML Models
catboost==1.2.2

# Utilities
python-dotenv==1.0.0
gdown==5.2.0
marshmallow==3.20.1
waitress==2.1.2
